requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow